refactor(tixlys): added nexus removed cqrs/meow
