+++ bins/auth/Cargo.toml
@@ -8,8 +8,7 @@ authors.workspace = true
 [dependencies]
 argon2 = { workspace = true }
 axum = { workspace = true, features = ["macros"] }
-cqrs = { path = "../../crates/cqrs" }
-meow_diator = { path = "../../crates/meow_diator" }
+nexus = { path = "../../crates/nexus" }
 password-hash = { workspace = true, features = ["rand_core", "getrandom", "alloc"] }
 pawz = { path = "../../crates/pawz" }
 serde = { workspace = true, features = ["derive"] }
