+++ bins/events/Cargo.toml
@@ -8,7 +8,6 @@ authors.workspace = true
 [dependencies]
 axum = { workspace = true }
 chrono = { workspace = true }
-cqrs = { path = "../../crates/cqrs" }
 sqlx = { workspace = true, features = ["sqlite", "migrate", "runtime-tokio-rustls"] }
 thiserror = { workspace = true }
 tokio = { workspace = true, features = ["full"] }
