--- crates/cqrs/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "cqrs"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-authors.workspace = true
-
-[dependencies]
-chrono = { workspace = true }
-thiserror = { workspace = true }
-tokio = { workspace = true, features = ["full"] }
-tower = { workspace = true, features = ["util"] }
-tracing = { workspace = true }
-ulid = { workspace = true }
-workspace-hack = { version = "0.1.0", path = "../workspace-hack" }
