--- crates/cqrs/src/lib.rs
@@ -1,2 +1,0 @@
-// mod bus;
-pub mod command;
