--- crates/meow_diator/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "meow_diator"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-authors.workspace = true
-
-[dependencies]
-thiserror = { workspace = true }
-tower = { workspace = true, features = ["util"] }
-workspace-hack = { version = "0.1.0", path = "../workspace-hack" }
-
-[dev-dependencies]
-tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
