+++ crates/nexus/src/command.rs
