+++ crates/workspace-hack/Cargo.toml
@@ -72,11 +72,13 @@ futures-sink = { version = "0.3" }
 getrandom = { version = "0.3", default-features = false, features = ["std"] }
 hyper = { version = "1", features = ["client", "http1", "http2", "server"] }
 hyper-util = { version = "0.1", features = ["client-legacy", "http1", "http2", "server", "service"] }
+libc = { version = "0.2" }
 tower = { version = "0.5", default-features = false, features = ["timeout"] }
 
 [target.x86_64-unknown-linux-gnu.build-dependencies]
 bitflags = { version = "2", default-features = false, features = ["std"] }
 futures-sink = { version = "0.3" }
 getrandom = { version = "0.3", default-features = false, features = ["std"] }
+libc = { version = "0.2" }
 
 ### END HAKARI SECTION
