+++ flake.nix
@@ -64,9 +64,8 @@
             fileset = lib.fileset.unions [
               ./Cargo.toml
               ./Cargo.lock
-              (craneLib.fileset.commonCargoSources ./crates/cqrs)
               (craneLib.fileset.commonCargoSources ./crates/pawz)
-              (craneLib.fileset.commonCargoSources ./crates/meow_diator)
+              (craneLib.fileset.commonCargoSources ./crates/nexus)
               (craneLib.fileset.commonCargoSources ./crates/workspace-hack)
               (craneLib.fileset.commonCargoSources crate)
             ];
